move reusable command options to a standalone file
