+++ src/util/fileCache.ts
@@ -1,6 +1,6 @@
 import bent from 'bent';
 import { Collection } from 'slash-create';
-import { ONE_HOUR } from './common';
+import { ONE_HOUR } from './constants';
 
 import { BRANCH, rawContentLink } from './linkBuilder';
 import { logger } from './logger';
